Because NszDecompressor decompresses all three major compressed file extensions: I renamed it to simply "Decompressor".
